Fix an int/long mismatch identified here: 	http://www.tortall.net/mu/blog/2005/12/01
Pointed out from SF #1365916.

Backport candidate.

+++ Python/bltinmodule.c
@@ -1908,8 +1908,9 @@ builtin_sorted(PyObject *self, PyObject *args, PyObject *kwds)
 	PyObject *newlist, *v, *seq, *compare=NULL, *keyfunc=NULL, *newargs;
 	PyObject *callable;
 	static const char *kwlist[] = {"iterable", "cmp", "key", "reverse", 0};
-	long reverse;
+	int reverse;
 
+	/* args 1-4 should match listsort in Objects/listobject.c */
 	if (!PyArg_ParseTupleAndKeywords(args, kwds, "O|OOi:sorted",
 		kwlist, &seq, &compare, &keyfunc, &reverse))
 		return NULL;
